手入力で都道府県を選べるようにしたよ
